+++ src/tests/test_cases/sections.py
@@ -215,3 +215,28 @@ def multiple_sections():
     My attention.
 
     """
+
+
+@expect(_D213)
+def false_positive_section_prefix():
+    """Toggle the gizmo.
+
+    Parameters
+    ----------
+    attributes_are_fun: attributes for the function.
+
+    """
+
+
+@expect(_D213)
+def section_names_as_parameter_names():
+    """Toggle the gizmo.
+
+    Parameters
+    ----------
+    notes : list
+        A list of wonderful notes.
+    examples: list
+        A list of horrible examples.
+
+    """